Add SizedText parser, deprecate Text & InvalidMac, work on mac parser & dispatcher
The Text parser has been superseeded by SizedText and ExactlySizedText,
and has been marked as deprecated. The original Text parser design fails
to map cleanly on the most common case, that of a single byte for length
before the actual text. SizedText expects up to a max capacity for its
content, whereas ExactlySizedText expectes a known size that is
specified in advance; no more, no less.

For the mac parser, we depend on crate macaddr available at
https://crates.io/crates/macaddr. Use MacAddr6 struct for normal
ethernet macs. Some Mac parser tests added.

The HtipError::InvalidMac is deprecated, any 6-byte slice is a valid
mac.

The dispatcher will eventually be able to call the appropriate parser,
depending on the type and the contents of the tlv to be parsed. The
parser constructors are registered to a type/prefix "key", so
appropriate parsers can be generated as needed.

Other changes: some From/TryFrom impls